requirements.txt
llama-index==0.9.33
openai==1.7.2
python-dotenv==1.0.0
